Create pull_request_template.ms
Title: [Identifier] Feature/Hotfix/Release/Task - Short title

# Context
[Link to Jira/Trello/GitHub issue]
Provide a description of the issue that’s being addressed. Aim to give enough information so that the reviewer doesn’t need to ask for additional details in order to provide a satisfactory review. 

# Approach
[Link to technical design document]
Provide a description of the technical approach to solve the issue. Consider getting a review on the approach before starting to code. 

# Changes
A list of changes that need to be made in order to merge the PR. Use it as a guide to enumerate all elements that must be completed.

# Commands
A list of commands that need to be run before/after merging the PR.

# Considerations
Additional information relevant to the PR. Are there side effects?